+++ packages/gatsby-plugin-emotion/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-emotion/compare/gatsby-plugin-emotion@4.0.1...gatsby-plugin-emotion@4.0.2) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-plugin-emotion
+
 <a name="4.0.1"></a>
 
 ## [4.0.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-emotion/compare/gatsby-plugin-emotion@4.0.0...gatsby-plugin-emotion@4.0.1) (2018-12-19)