+++ packages/gatsby-remark-smartypants/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-smartypants/compare/gatsby-remark-smartypants@2.0.7...gatsby-remark-smartypants@2.0.8) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-remark-smartypants
+
 <a name="2.0.7"></a>
 
 ## [2.0.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-smartypants/compare/gatsby-remark-smartypants@2.0.6...gatsby-remark-smartypants@2.0.7) (2018-11-29)