+++ packages/gatsby-source-graphql/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-graphql/compare/gatsby-source-graphql@2.0.9...gatsby-source-graphql@2.0.10) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-source-graphql
+
+## [2.0.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-graphql/compare/gatsby-source-graphql@2.0.8...gatsby-source-graphql@2.0.9) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-source-graphql
+
 <a name="2.0.8"></a>
 
 ## [2.0.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-graphql/compare/gatsby-source-graphql@2.0.7...gatsby-source-graphql@2.0.8) (2018-12-29)