+++ packages/gatsby-transformer-javascript-static-exports/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-javascript-static-exports/compare/gatsby-transformer-javascript-static-exports@2.1.4...gatsby-transformer-javascript-static-exports@2.1.5) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-transformer-javascript-static-exports
+
 <a name="2.1.4"></a>
 
 ## [2.1.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-javascript-static-exports/compare/gatsby-transformer-javascript-static-exports@2.1.3...gatsby-transformer-javascript-static-exports@2.1.4) (2018-11-29)